% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rboxplot.R
\name{rboxplot}
\alias{rboxplot}
\title{Randomly sample a set of 4 boxplots for interpretation testing}
\usage{
rboxplot(n = 100, mu = 0.6, sd = 1.5, skew = 10)
}
\arguments{
\item{n}{the sample size per group, per boxplot}

\item{mu}{the difference in center}

\item{sd}{the difference in spread}

\item{skew}{the difference in skewness}
}
\value{
a `data.frame` suitable for plotting with `ggplot2`
}
\description{
Randomly sample a set of 4 boxplots for interpretation testing
}
